+++ e2e-tests-typescript/src/step_definitions/Topics.steps.ts
@@ -1,6 +1,7 @@
 import { Given, When, Then, setDefaultTimeout } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { fixture } from "../hooks/pageFixture";
+import { expectVisibility, ensureCheckboxState } from "../services/uiHelper";
 
 setDefaultTimeout(60 * 1000 * 2);
 
@@ -20,6 +21,7 @@ Given('Topics AddATopic visible', async () => {
 
 Given('Topics DeleteSelectedTopics active is: {string}', async (state: string) => {
   const isEnabled = await fixture.topics.topicDeleteSelectedTopicsButton().isEnabled();
+
   expect(isEnabled.toString()).toBe(state);
 });
 
@@ -37,31 +39,14 @@ Given('Topics PurgeMessagesOfSelectedTopics active is: {string}', async (state: 
   expect(isEnabled.toString()).toBe(state);
 });
 
-When('Topic SelectAllTopic visible is: {string}', async (state: string) => {
-  const checkbox = fixture.topics.topicSelectAllCheckBox();
-  const shouldBeVisible = state === "true";
-
-  if (shouldBeVisible) {
-    await expect(checkbox).toBeVisible();
-  } else {
-    await expect(checkbox).toHaveCount(0);
-  }
+When('Topic SelectAllTopic visible is: {string}', async (visible: string) => {
+  await expectVisibility(fixture.topics.topicSelectAllCheckBox(), visible)
 });
 
 
 Then('Topic SelectAllTopic checked is: {string}', async (state: string) => {
   const checkbox = fixture.topics.topicSelectAllCheckBox();
-  const desiredState = state === "true";
-  const currentState = await checkbox.isChecked();
-
-  if (currentState !== desiredState) {
-    if (desiredState) {
-      await checkbox.check();
-    } else {
-      await checkbox.uncheck();
-    }
-  }
-
+  await ensureCheckboxState(checkbox, state);
   const actual = await checkbox.isChecked();
   expect(actual.toString()).toBe(state);
 });
@@ -76,52 +61,26 @@ When('Topics serchfield input {string}', async (topicName: string) => {
 });
 
 Then('Topic named: {string} visible is: {string}', async (topicName: string, visible: string) => {
-    const topic = fixture.topics.topicNameLink(topicName);
-    if (visible === "true") {
-    await expect(topic).toBeVisible();
-  } else if (visible === "false") {
-    await expect(topic).toHaveCount(0);
-  } else {
-    throw new Error(`Invalid visibility value: ${visible}`);
-  }
+  await expectVisibility(fixture.topics.topicNameLink(topicName), visible);
 });
 
 When('Topic serchfield input cleared', async () => {
-    const textBox = fixture.topics.topicSearchField();
+  const textBox = fixture.topics.topicSearchField();
 
-    await textBox.fill('');
+  await textBox.fill('');
 
   const text = await textBox.inputValue();
   expect(text).toBe('');
 });
 
 When('Topics ShowInternalTopics switched is: {string}', async (state: string) => {
-    const checkBox = fixture.topics.topicShowInternalTopics();
-    const desiredState = state === "true";
-    const isChecked = await checkBox.isChecked();
+  const checkBox = fixture.topics.topicShowInternalTopics();
 
-  if (isChecked !== desiredState) {
-    if (desiredState) {
-      await checkBox.check();    // turn ON
-    } else {
-      await checkBox.uncheck();  // turn OFF
-    }
-  }
+  await ensureCheckboxState(checkBox, state);
 });
 
 When('Topic row named: {string} checked is: {string}', async (topicName: string, state: string) => {
   const checkbox = fixture.topics.topicRowCheckBox(topicName);
-  const desiredState = state === 'true';
-  const isChecked = await checkbox.isChecked();
-
-  if (isChecked !== desiredState) {
-    if (desiredState) {
-      await checkbox.check();
-    } else {
-      await checkbox.uncheck();
-    }
-  }
-
-  const finalState = await checkbox.isChecked();
-  expect(finalState.toString()).toBe(state);
+  
+  await ensureCheckboxState(checkbox, state);
 });
