+++ e2e-tests-typescript/src/step_definitions/TopicsCreate.steps.ts
@@ -1,8 +1,8 @@
 import { Given, When, Then, setDefaultTimeout } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { fixture } from "../hooks/pageFixture";
-import { expectVisibility } from "../services/visibilityHelper";
-import { CustomWorld } from '../support/CustomWorld';
+import { expectVisibility } from "../services/uiHelper";
+import { CustomWorld } from "../support/customWorld";
 import { generateName } from "../services/commonFunctions";
 
 setDefaultTimeout(60 * 1000 * 2);
@@ -107,19 +107,13 @@ When('TopicCreate Create topic clicked', async function (this: CustomWorld) {
 Then('Header starts with: {string}', async function (this: CustomWorld, prefix: string) {
   const topicName = this.getValue<string>(`topicName-${prefix}`);
   const header = fixture.page.getByRole('heading', { name: topicName });
+
   await expect(header).toBeVisible();
 });
 
 Then('Topic name started with: {string} visible is: {string}', async function (this: CustomWorld, prefix: string, visible: string) {
   const topicName = this.getValue<string>(`topicName-${prefix}`);
-  const locator = fixture.topics.topicNameLink(topicName);
-
-  const shouldBeVisible = visible === 'true';
-  if (shouldBeVisible) {
-    await expect(locator).toBeVisible();
-  } else {
-    await expect(locator).toHaveCount(0);
-  }
+  await expectVisibility(fixture.topics.topicNameLink(topicName), visible);
 });
 
 Then('TopicCreate TimeToRetainData value is: {string}', async (expectedValue: string) => {
